requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
emoji==1.2.0
ephem==3.7.7.1
PySocks==1.7.1
python-telegram-bot==13.5
pytz==2021.1
six==1.16.0
tornado==6.1
tzlocal==2.1
python-dotenv==0.19.1
clarifai-grpc==7.11.0